fast end cleaning
