+++ app/src/components/AvatarLock.styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: 8,
+  },
+  lockIcon: {
+    // Size is set dynamically based on UIConfig.progressSection.iconSize
+  },
+})
+
